validation v1 version
thymeleaf class & if 문법을 통한 예외처리 형식 적용

+++ validation-start/src/main/java/hello/itemservice/domain/item/Item.java
@@ -7,13 +7,13 @@ public class Item {
 
     private Long id;
     private String itemName;
-    private Integer price;
+    private Long price;
     private Integer quantity;
 
     public Item() {
     }
 
-    public Item(String itemName, Integer price, Integer quantity) {
+    public Item(String itemName, Long price, Integer quantity) {
         this.itemName = itemName;
         this.price = price;
         this.quantity = quantity;
